Makes generating passwords from wordlists possible.
This is still kinda experimental, but for now you can already try it.
The `--wordlist` CLI option allows for adding a new wordlist, and the
syntax for referencing it is `\w[name]`. Example:

    passgen --wordlist german:/usr/share/dict/ngerman "\w[german]"

I will have to do some more work to clean up the code, make it more
efficient, and more importantly adding markov chain generation back into
the code.

To that end, I'm planning to implement a hash-based radix tree to
efficiently store wordlists in-memory, in a way that handles unicode
characters.

Aside from that, other tasks that need to be done include cleaning up
the mess that is the current data structure implementations, documenting
things some more, and perhaps adding some kind of option to have a
config file with presets, looking something like

    wordlist english /usr/share/dict/english
    wordlist german /usr/share/dict/ngerman
    preset mypass \w[german]
    preset other [a-z]{26}

Finally, I want to implement the ability to more precisely control how
the length is chosen with length modifiers such as "[a-z]{25,35}".
Right now, it just picks any one of the possibilities, as such the
distribution of outcomes is linear. I want to add the ability to specify
"[a-z]{binomial(26, 2.0)}" to achieve a binomial distribution there.
This is definitely over-engineered, but it matches real-world length
data more accurately, I think.