Adds tests for the `passgen` binary.
This commit adds a script, `scripts/ci-test-passgen.sh` that runs passgen a few times, with different inputs and options. This script is called in the CI after the build steps to ensure that the resulting binary works as it should and nothing is inadvertently broken. This is especially useful to ensure that the seccomp filters are working as expected.