Fixes "detected dubious ownership" git error in CI.
The CI started to fail builds with the error message:

    fatal: detected dubious ownership in repository at
    '/builds/xfbs/passgen'

Tracking it down, it looks to be an issue with an older version of Git
(only jobs that use Git 2.30.2 have this issues, other jobs are not
affected). For that reason, building Git info is disabled in all test
jobs, and all build jobs that use the affected version of Git will fix
it by running

    git config --global --add safe.directory '*'

This has the effect of trusting all directories, which makes sense for
the CI. It is possible to simplify this by adding this command into the
Dockerfiles.