Fixes hashmap unicode bug: improper equal comparison
Fixes a bug in passgen_hashmap where using hashmaps with raw unicode
codepoints (using passgen_hashmap_context_unicode), keys wouldd not be
compared fully. The issue was that in the unicode_equal() function, it
would compute the lengths of the two unicode strings:

  size_t rhs_len = unicode_len(rhs);
  size_t lhs_len = unicode_len(lhs);

And then later use memcmp() to compare them if they match:

  return 0 == memcmp(lhs, rhs, rhs_len);

The issue here is that the unicode_len() function returns the string
length in terms of codepoints, whereas memcmp() expects a byte length.
The fix here is to multiply the length by sizeof(uint32_t), like such:

  return 0 == memcmp(lhs, rhs, rhs_len * sizeof(uint32_t));

Also added more testing to make sure this works as expected.