Adds `msan`, `usan` and `asan` targets to Makefile.
LLVM Sanitizers are immensely helpful in tracking down bugs in C and C++
codebases. They instrument the code at compile time, so that it's
dynamic behaviour can be analyzed at runtime to make sure it does not do
anything bad (such as access memory out of bounds, trigger undefined
behaviour, or have memory leaks).

While the CI pipeline is already running these, exposing them in the
Makefile makes it easier to run them locally. Simply run one of:

  make asan
  make usan
  CC=clang make msan

This uses the same `sanitizer.sh` script that the CI runs. Some issues
do not result in a failure, it is important to scroll through the output
and see if there are any warnings or errors generated.