Moves entropy out of passgen_env.
This removes the `find_entropy` and `entropy` fields of `passgen_env`,
and insteads adds a `double *entropy` parameter to the generation
methods. When passed a non-NULL value, the entropy will be calculated in
this variable.

This change allows the `passgen_env` to be used concurrently by multiple
generation threads.

Modifies CLI and unit tests to make use of the new mechanism.